fastapi-backend/requirements.txt
boto3
python-dotenv
fastapi
uvicorn
mysql-connector-python